Separando a lista vertical de personagens em um widget separadao
