+++ lib/main.dart
@@ -16,6 +16,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
+        fontFamily: "Quantico",
         primarySwatch: primaryBlack,
       ),
       initialRoute: NavigationRoute.home.name,
