+++ lib/pages/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:super_heroes_landing/widgets/vertical_list_view.dart';
 
 import '../models/results.dart';
 import '../repository/character_repository.dart';
@@ -11,90 +12,14 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  late List<Results> charactersList = [];
-  final scrollController = ScrollController();
-  bool hasMoreCharacters = true;
-  int totalCharacters = 0;
-  int offset = 25;
-
-  @override
-  void initState() {
-    super.initState();
-
-    CharacterRepository().fetchCharacters().then((response) => {
-          charactersList.addAll(response.requestdata.results),
-          totalCharacters = response.requestdata.total
-        });
-
-    scrollController.addListener(() {
-      if (scrollController.position.maxScrollExtent ==
-          scrollController.offset) {
-        if (charactersList.length == totalCharacters) {
-          hasMoreCharacters = false;
-        }
-
-        if (charactersList.length < totalCharacters) {
-          CharacterRepository()
-              .fetchCharacters(offset)
-              .then((response) => setState(() {
-                    charactersList.addAll(response.requestdata.results);
-                    offset += 25;
-                  }));
-        }
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    scrollController.dispose();
-
-    super.dispose();
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text(
-          "Super Hero Landing",
-          style: TextStyle(fontFamily: "Quantico"),
-        ),
+        title: const Text("Super Hero Landing"),
         centerTitle: true,
       ),
-      body: ListView.builder(
-          controller: scrollController,
-          itemCount: charactersList.length + 1,
-          itemBuilder: (context, index) {
-            if (index < charactersList.length) {
-              return Center(
-                child: Column(
-                  children: [
-                    Text(
-                      index.toString(),
-                      style: const TextStyle(
-                        fontFamily: "Quantico",
-                      ),
-                    ),
-                    Text(
-                      charactersList[index].name,
-                      style: const TextStyle(
-                          fontWeight: FontWeight.bold, fontFamily: "Quantico"),
-                    ),
-                  ],
-                ),
-              );
-            } else {
-              return Padding(
-                padding: const EdgeInsets.symmetric(vertical: 32),
-                child: Center(
-                  child: hasMoreCharacters
-                      ? const CircularProgressIndicator()
-                      : const Text("Sem mais dados para carregar"),
-                ),
-              );
-            }
-          }),
+      body: const VerticalListView(),
     );
   }
 }
